+++ CHANGELOG.md
@@ -37,6 +37,7 @@ For experimental package changes, see the [experimental CHANGELOG](experimental/
 
 * feat(sdk-metrics): add constructor option to add metric readers [#4427](https://github.com/open-telemetry/opentelemetry-js/pull/4427) @pichlermarc
   * deprecates `MeterProvider.addMetricReader()` please use the constructor option `readers` instead.
+* chore: Semantic Conventions export individual strings [4185](https://github.com/open-telemetry/opentelemetry-js/issues/4185)
 
 ### :bug: (Bug Fix)